backend/requirements.txt
fastapi==0.116.1
uvicorn==0.30.6
gunicorn==23.0.0
openai-whisper==20231117
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
requests==2.32.4
httpx==0.28.1
tenacity==9.1.2
structlog==25.4.0
prometheus-client==0.22.1
SQLAlchemy==2.0.43
pymysql==1.1.1
alembic==1.13.2
python-jose==3.5.0
redis==6.4.0
cryptography==45.0.6
opentelemetry-sdk==1.36.0
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-instrumentation-requests==0.57b0
opentelemetry-instrumentation-sqlalchemy==0.57b0
opentelemetry-instrumentation-pika==0.57b0
presidio-analyzer==2.2.354  ; python_version>='3.10'
PyJWT==2.9.0

